Info security

+++ server/common/controllers/department.controller.ts
@@ -4,11 +4,26 @@ import { user } from "../middlewares/user.middleware.js";
 import { logging } from "../../engine/logging.js";
 
 export const departmentController = {
+  /**
+   * 
+   * @param req 
+   * @param res 
+   */
   addDepartment: (req: Request, res: Response) => {
+    const reqId = req.user.userId;
     const payload = req.body;
+    const log = logging.userLogs(payload.superuserId);
+
+    // Another superuser can't create a department on another superuser other than their own.
+    if (reqId !== payload.superuserId) {
+      log.warn(`User ${reqId} tried to create a department on $(payload.superuserId)`);
+      return res.status(500).json({
+        status: false,
+        error: `User ${reqId} don't have the required permission to perform the operation.`,
+      });
+    }
     departmentCrud.addDept(Object.assign(payload))
       .then((dept) => {
-        const log = logging.userLogs(String(dept.superuserId));
         log.info(
           `Department ${dept.departmentId} has just been created by  User ${req.user.userId}`)
         return res.status(201).json({
@@ -16,7 +31,7 @@ export const departmentController = {
         });
       })
       .catch((err) => {
-        logging.controllerLogger.error(new Error(err));
+        log.error(new Error(err));
         return res.status(500).json({
           status: false,
           error: err,
@@ -59,7 +74,9 @@ export const departmentController = {
       });
   },
 
-  getDept: (req: Request, res: Response) => {
+  getDept: async (req: Request, res: Response) => {
+    const reqId = req.user.userId;
+    const _superuserId = await user._user_id(reqId);
     const {
       params: { deptId }
     } = req;
@@ -67,6 +84,17 @@ export const departmentController = {
     departmentCrud.findDept({ departmentId: deptId })
       .then((dept) => {
         const log = logging.userLogs(String(dept?.superuserId));
+
+        // Only user assigned to task has access to the task
+        if (_superuserId !== dept?.superuserId) {
+          log.warn(
+            `User ${reqId} tried to access an unauthorized department ${dept?.departmentId}`
+          );
+          return res.status(500).json({
+            status: false,
+            error: `User ${reqId} doesn't have the required permission to access this information`,
+          })
+        }
         log.info(`Department: ${deptId} retrieved by User: ${req.user.userId}`)
         return res.status(200).json({
           status: true,
@@ -82,18 +110,32 @@ export const departmentController = {
       })
   },
 
-  updateDept: (req: Request, res: Response) => {
+  updateDept: async (req: Request, res: Response) => {
+    const reqId = req.user.userId;
+    const reqAdmin = await user._user_id(reqId);
     const {
       params: { deptId },
       body: payload,
     } = req;
+
+    const dept = await departmentCrud.findDept({ departmentId: deptId });
+    const log = logging.userLogs(String(dept?.superuserId));
+
+    // Reject update if the superuser is of another superuser server
+    if (reqAdmin !== dept?.superuserId) {
+      log.warn(`User ${reqId} tried to update department ${deptId}`);
+      return res.status(500).json({
+        status: false,
+        error: `User ${reqId} don't have the required permission to perform this operation.`,
+      });
+    }
 
     // If the payload doesn't have any keys, return error
     if (!Object.keys(payload).length) {
       return res.status(400).json({
         status: false,
         error: {
-          message: 'Body is empty, hence cannot update the dept',
+          message: 'no update provided',
         }
       })
     }
@@ -102,7 +144,6 @@ export const departmentController = {
         return departmentCrud.findDept({ departmentId: deptId });
       })
       .then((dept) => {
-        const log = logging.userLogs(String(dept?.superuserId))
         log.info(
           `Update on Department ${dept?.departmentId} performed by ${req.user.userId}`
         );
@@ -112,7 +153,7 @@ export const departmentController = {
         });
       })
       .catch((err) => {
-        logging.controllerLogger.error(new Error(err));
+        log.error(new Error(err));
         return res.status(500).json({
           status: false,
           error: err,
@@ -121,11 +162,22 @@ export const departmentController = {
   },
 
   deleteDept: async (req: Request, res: Response) => {
+    const reqId = req.user.userId;
     const {
       params: { deptId }
     } = req;
+
     const dept = await departmentCrud.findDept({ deptId: deptId })
     const log = logging.userLogs(String(dept?.superuserId));
+
+    // Only the superuser who created can possibly destroy.
+    if (reqId !== dept?.superuserId) {
+      log.warn(`User ${reqId} tried to delete department ${deptId}`);
+      return res.status(500).json({
+        status: false,
+        error: `User ${reqId} does not have the required permission to perform this operation.`,
+      });
+    }
     departmentCrud.deleteDept({ departmentId: deptId })
       .then((numberOfDepartmentDeleted) => {
         log.warn(`Department ${deptId} was deleted by user ${req.user.userId}`
