+++ server/common/controllers/task.controller.ts
@@ -2,13 +2,50 @@ import { Request, Response } from 'express';
 import { taskCrud } from '../models/task.js';
 import { user } from '../middlewares/user.middleware.js';
 import { logging } from '../../engine/logging.js';
+import { userCrud } from '../models/user.js';
+import { departmentCrud } from '../models/department.js';
 
 export const taskController = {
-  newTask: (req: Request, res: Response) => {
+  /**
+   * 
+   * @param req 
+   * @param res 
+   */
+  newTask: async (req: Request, res: Response) => {
+    const reqId = req.user.userId;
     const payload = req.body;
+    const log = logging.userLogs(payload.superuserId);
+
+    // Confirm payload info
+    const _userId = payload.userId;
+    const _departmentId = payload.departmentId
+    const _user = await userCrud.findUser({ userId: _userId });
+    const _department = await departmentCrud.findDept({ departmentId: _departmentId });
+  
+    // confirm payload info
+    if (reqId !== payload.superuserId) {
+      log.error(`User ${reqId} tried to create a task on ${payload.superuserId}`);
+      return res.status(500).json({
+        status: false,
+        error: `User ${reqId} don't have the required permission to perform this operation`,
+      });
+    }
+    if (reqId !== _user?.superuserId) {
+      log.error(`User ${_userId} does not exist in superuser ${payload.superuserId}`)
+      return res.status(500).json({
+        status: false,
+        error: `User ${_userId} does not exist`,
+      })
+    }
+    if (reqId !== _department?.superuserId) {
+      log.error(`Department ${_departmentId} does not exist in superuser ${payload.superuserId}`)
+      return res.status(500).json({
+        status: false,
+        error: `Department ${_departmentId} does not exist`,
+      })
+    }
     taskCrud.createTask(Object.assign(payload))
       .then((task) => {
-        const log = logging.userLogs(String(task.superuserId));
         log.info(`Task: ${task.taskId} has just been created by User: ${req.user.userId}`);
         return res.status(201).json({
           status: true,
@@ -16,7 +53,7 @@ export const taskController = {
         });
       })
       .catch((err) => {
-        logging.controllerLogger.error(new Error(err));
+        log.error(new Error(err));
         return res.status(500).json({
           status: false,
           error: err,
@@ -24,8 +61,9 @@ export const taskController = {
       });
   },
 
-  getTask: (req: Request, res: Response) => {
+  getTask: async (req: Request, res: Response) => {
     const reqId = req.user.userId;
+    const _superuser = await user._user_id(reqId);
     const {
       params: { taskId }
     } = req;
@@ -33,13 +71,15 @@ export const taskController = {
     taskCrud.findTask({ taskId: taskId })
       .then((task) => {
         const log = logging.userLogs(String(task?.superuserId));
-        if (reqId !== task?.userId) {
+
+        // Only user assigned to task has access to the task
+        if (_superuser !== task?.superuserId) {
           log.warn(
             `User ${reqId} tried to access an unauthorized task ${task?.taskId}`
           );
           return res.status(500).json({
             status: false,
-            error: `User ${reqId} does not have the required permission`,
+            error: `User ${reqId} doesn't have the required permission to access this information`,
           });
         }
         log.info(`Task: ${task?.taskId}, accessed by User ${reqId}`);
@@ -57,7 +97,7 @@ export const taskController = {
       });
   },
 
-  getAllTasks: (req: any, res: Response) => {
+  getAllTasks: (req: Request, res: Response) => {
     const {
       user: { userId }
     } = req;
@@ -91,6 +131,8 @@ export const taskController = {
   },
 
   updateTask: async (req: Request, res: Response) => {
+    const reqId = req.user.userId;
+    const reqAdmin = await user._user_id(reqId);
     const {
       params: { taskId },
       body: payload
@@ -98,6 +140,15 @@ export const taskController = {
 
     const task = await taskCrud.findTask({ taskId: taskId});
     const log = logging.userLogs(String(task?.superuserId));
+    
+    // Reject update if the superuser or admin is from another superuser server
+    if (reqAdmin !== task?.superuserId) {
+      log.warn(`An unknown user ${reqId} tried to update task ${taskId}`);
+      return res.status(500).json({
+        status: false,
+        error: `User ${reqId} don't have the required permission to perform this operation.`,
+      });
+    }
 
     // if the payload does not have any keys, return error
     if (!Object.keys(payload).length) {
@@ -105,7 +156,7 @@ export const taskController = {
       return res.status(400).json({
         status: false,
         error: {
-          message: 'Body is empty, hence cannot update the user',
+          message: 'No update provided',
         },
       });
     }
@@ -130,13 +181,22 @@ export const taskController = {
   },
 
   deleteTask: async (req: Request, res: Response) => {
+    const reqId = req.user.userId;
     const {
       params: { taskId }
     } = req;
 
     const task = await taskCrud.findTask({ taskId: taskId })
-    const _superuserId = await user._user_id(String(task?.userId));
-    const log = logging.userLogs(String(_superuserId));
+    const log = logging.userLogs(String(task?.superuserId));
+
+    // Only superuser who created the task can delete it.
+    if (reqId !== task?.superuserId) {
+      log.warn(`User ${reqId} tried to delete task ${taskId}.`);
+      return res.status(500).json({
+        status: false,
+        error: `User ${reqId} does not have the required permission to perform this operation`,
+      });
+    }
 
     taskCrud.deleteTask({ taskId: taskId })
       .then((numberOfTasksDeleted) => {
